Move `partition`
