+++ src/index.js
@@ -75,6 +75,7 @@ module.exports = {
     utils: {
       arrayify: require("./utils/arrayify.js"),
       getLast: require("./utils/get-last.js"),
+      partition: require("./utils/partition.js"),
       isNonEmptyArray: require("./common/util.js").isNonEmptyArray,
     },
   },
